I am attaching the first part of the homework.
Task 1 - Create a function which returns max even element in array from array of strings

Task 2 - Create a function which finds unique elements in array

Task 3 - Write a function – calculateSum. It should accept an array of numers and count the sum of this numbers.

Task 4 - Implement the function calculateNumber. Given an array of ones and zeroes. You need to convert the equivalent binary value to an integer