Adding the second part of the homework.
Task 1 - Write a function that accepts a string and outputs statistics: the number of letters, the number of digits and the number of other characters.

Task 2 - Write a function that converts CSS style names with a hyphen into a name in Сamelcase style.

Task 3 - Write a function that accepts a phrase and turns it into an abbreviation.

Task 4 - Write a function that takes a string and a delimiter and returns an array of substrings split by the specified delimiter.

Task 5 - Write a function that receives a url and displays detailed information about it.